Make Toolbar consistent.
The toolbar currently is context sensitive and many parts only show up when at least one object is selected.
Hiding objects will change the ui that way and it may happen that outliner and other important buttons can dissapear. Now the UI will always be drawn. Even when objects are hidden or deleted.